# api/openapi.yaml
openapi: 3.0.3
info:
  title: Spillway admin API
  version: 0.4.0
  description: |
    The API behind the dashboard. Its only caller is the dashboard's server-side layer, which signs users in
    with POST /admin/login and forwards the returned token as `Authorization: Bearer <token>`. The browser
    never sees the token. TypeScript types for web/ are generated from this file, and a Go test checks that
    the real responses match it.

    Each operation says which roles may call it with `x-access`: `public`, `viewer` (any signed-in user)
    or `admin`. The Go router registers the same access level, and a test fails if they disagree.
paths:
  /admin/login:
    post:
      operationId: login
      summary: Sign in with a username and password
      x-access: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Signed in
          headers:
            Cache-Control: { schema: { type: string }, description: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400":
          description: The body was not a username and a password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: The username or password is wrong. The answer is the same for an unknown user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Five failed sign-ins in a minute for this address or username
          headers:
            Retry-After: { schema: { type: integer, minimum: 1 }, description: Seconds until the next try is allowed }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/me:
    get:
      operationId: me
      summary: The signed-in user and their role
      x-access: viewer
      security: [{ bearer: [] }]
      responses:
        "200":
          description: The user the token belongs to
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: No token, a bad token, or an expired one (`session_expired`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/status:
    get:
      operationId: status
      summary: Whether Spillway's dependencies are reachable
      x-access: viewer
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Dependency status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "401":
          description: No token, a bad token, or an expired one
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
    Role:
      type: string
      enum: [admin, viewer]
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
    User:
      type: object
      required: [username, role]
      properties:
        username: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    LoginResponse:
      type: object
      required: [token, expires_at, user]
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
...
    Status:
      type: object
      required: [postgres, redis]
      properties:
        postgres: { type: string, enum: [up, down, unknown] }
        redis: { type: string, enum: [up, down, disabled], description: "disabled means REDIS_URL is not set" }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...